+++ src/main/java/lotto/model/LottoMoney.java
@@ -6,10 +6,10 @@ public class LottoMoney {
     private static final String ERROR_NOT_POSITIVE = "구입 금액은 양수여야 합니다.";
     private static final String ERROR_NOT_SUFFICIENT_TO_BUY = "로또 구입 금액이 부족합니다.";
 
-    private final int lottoMoney;
+    private final long lottoMoney;
     private final int numberOfManualLottos;
 
-    public LottoMoney(int lottoMoney, int numberOfManualLottos) {
+    public LottoMoney(long lottoMoney, int numberOfManualLottos) {
         validatePositive(lottoMoney);
         validateUnitPrice(lottoMoney);
         validateCanBuyLottos(lottoMoney, numberOfManualLottos);
@@ -17,26 +17,26 @@ public class LottoMoney {
         this.numberOfManualLottos = numberOfManualLottos;
     }
 
-    private void validatePositive(int lottoMoney) {
+    private void validatePositive(long lottoMoney) {
         if (lottoMoney <= 0) {
             throw new IllegalArgumentException(ERROR_NOT_POSITIVE);
         }
     }
 
-    private void validateUnitPrice(int lottoMoney) {
+    private void validateUnitPrice(long lottoMoney) {
         if (lottoMoney % LOTTO_PRICE != 0) {
             throw new IllegalArgumentException(ERROR_NOT_DIVIDED_BY_UNIT_PRICE);
         }
     }
 
-    private void validateCanBuyLottos(int lottoMoney, int numberOfManualLottos) {
-        if (lottoMoney - numberOfManualLottos * LOTTO_PRICE < 0) {
+    private void validateCanBuyLottos(long lottoMoney, int numberOfManualLottos) {
+        if (lottoMoney - (long)numberOfManualLottos * LOTTO_PRICE < 0) {
             throw new IllegalArgumentException(ERROR_NOT_SUFFICIENT_TO_BUY);
         }
     }
 
     int getAutoLottoSize() {
-        return lottoMoney / LOTTO_PRICE - numberOfManualLottos;
+        return Math.toIntExact(lottoMoney / LOTTO_PRICE - numberOfManualLottos);
     }
 
     public long getLottoMoney() {
